Seperation of Concerns:   move functions of game board to eventsgame.js from events.js   begin AJAX call to update game.cells to api.js Update font type for game-board element to index.scss
