+++ assets/scripts/winlogic.js
@@ -1,0 +1,13 @@
+'use strict'
+
+// let horizontalWin => () {
+//   if (cellIndex)
+// }
+
+// var winsRowX = function() {
+//   return allThreeX(cellValue('a'), cellValue('b'), cellValue('c')) ||
+//           allThreeX(cellValue('d'), cellValue('e'), cellValue('f')) ||
+//           allThreeX(cellValue('g'), cellValue('h'), cellValue('i'));
+// }
+// var allThreeX = function(cellOne, cellTwo, cellThree) {
+// }
