+++ assets/styles/index.scss
@@ -27,6 +27,8 @@ p {
 .game-cell {
   height: $height;
   width: $width;
+  color: $crazy-yellow;
+  font-size: 36px;
 }
 
 .top-row {
